Refactoring for general read file function
